Use appropriate cast operators

+++ Hazel/src/Hazel/ImGui/ImGuiLayer.cpp
@@ -63,7 +63,10 @@ namespace Hazel
 	{
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::Get();
-		io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+		io.DisplaySize = ImVec2(
+			static_cast<float>(app.GetWindow().GetWidth()), 
+			static_cast<float>(app.GetWindow().GetHeight())
+		);
 
 		float time = static_cast<float>(glfwGetTime());
 		io.DeltaTime = m_Time > 0.0 ? (time - m_Time) : (1.0f / 60.0f);
@@ -150,7 +153,7 @@ namespace Hazel
 		ImGuiIO& io = ImGui::GetIO();
 		int keycode = e.GetKeyCode();
 		if (keycode > 0 && keycode < 0x10000)
-			io.AddInputCharacter((unsigned short)keycode);
+			io.AddInputCharacter(static_cast<unsigned short>(keycode));
 
 		return false;
 	}
@@ -158,7 +161,10 @@ namespace Hazel
 	bool ImGuiLayer::OnWindowResizeEvent(WindowResizeEvent& e)
 	{
 		ImGuiIO& io = ImGui::GetIO();
-		io.DisplaySize = ImVec2(e.GetWidth(), e.GetHeight());
+		io.DisplaySize = ImVec2(
+			static_cast<float>(e.GetWidth()), 
+			static_cast<float>(e.GetHeight())
+		);
 		io.DisplayFramebufferScale = ImVec2(1.0f, 1.0f);
 		glViewport(0, 0, e.GetWidth(), e.GetHeight());
